examples/requirements.txt
datasets
transformers==4.45.0
scikit-learn
Pillow
urllib3
